feat(proto): inherent Decimal::to_decimal() method (closes #117)
Add proto_ext.rs with impl proto::Decimal { to_decimal() } as an ergonomic
alias for the existing decimal_from_proto free function. Tests cover positive,
negative (sign-extension path), zero, and parity with the free function.
